Autoscale progress bar
Textual's CSS is, I think, still a bit buggy. Or else counterintuitive. The min-width and max-width properties have different ways of taking margin and padding into account. Also, combining max-width: something with width 1fr does not result in the latter filling the rest of the space. At least not for the progress bar.